pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.11.0.86",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "pyzbar>=0.1.9",
    "qrcode>=8.0",
    "reportlab>=4.2.5",
    "streamlit>=1.41.1",
    "twilio>=9.4.4",
]
